refactor: feature/#50 BoardService to BoardServiceImpl
